TF-TRT: Split fixture interfaces into files.
Factor out test fixture interfaces in convert_nodes_test.cc into
dedicated files under convert/fixtures, improving code clarity and
readability.
Factor out GPU Tensor creation utilities into dedicate tensor_testutil.h to
mirror tensorflow/framework/tensor_testutil.h.
